# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Used-car marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Dealer {
  dealer_id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postal_code?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
}

export interface Listing {
  id: string;
  vin: string;
  /** @nullable */
  dealer_id?: string | null;
  /** @nullable */
  stock_number?: string | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  make?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  model_number?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  transmission?: string | null;
  /** @nullable */
  series?: string | null;
  /** @nullable */
  series_detail?: string | null;
  /** @nullable */
  door_count?: number | null;
  /** @nullable */
  odometer?: number | null;
  /** @nullable */
  engine_cylinders?: number | null;
  /** @nullable */
  engine_displacement?: string | null;
  /** @nullable */
  engine?: string | null;
  /** @nullable */
  drivetrain?: string | null;
  /** @nullable */
  exterior_color?: string | null;
  /** @nullable */
  interior_color?: string | null;
  /** @nullable */
  msrp?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  certified?: boolean | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  features?: string | null;
  photo_urls?: string[];
  /** @nullable */
  city_mpg?: number | null;
  /** @nullable */
  highway_mpg?: number | null;
  /** @nullable */
  vehicle_detail_link?: string | null;
  /** @nullable */
  inventory_date?: string | null;
  /** @nullable */
  photos_last_modified?: string | null;
  /** @nullable */
  age?: number | null;
  status: string;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  updated_at?: string | null;
  dealer?: Dealer;
}

export interface ListingsPage {
  items: Listing[];
  total: number;
  limit: number;
  offset: number;
}

export interface ListingFilters {
  makes: string[];
  models: string[];
  /** @nullable */
  year_min?: number | null;
  /** @nullable */
  year_max?: number | null;
  /** @nullable */
  price_min?: number | null;
  /** @nullable */
  price_max?: number | null;
  /** @nullable */
  odometer_max?: number | null;
}

export interface ListingsStats {
  total: number;
  certified_count?: number;
  /** @nullable */
  avg_price?: number | null;
  /** @nullable */
  min_price?: number | null;
  /** @nullable */
  max_price?: number | null;
  makes_count?: number;
}

export type LeadInputLeadType = typeof LeadInputLeadType[keyof typeof LeadInputLeadType];


export const LeadInputLeadType = {
  inquiry: 'inquiry',
  click_for_price: 'click_for_price',
} as const;

export interface LeadInput {
  /** @nullable */
  vin?: string | null;
  lead_type: LeadInputLeadType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  website?: string | null;
}

export interface Lead {
  id: string;
  /** @nullable */
  vin?: string | null;
  /** @nullable */
  listing_id?: string | null;
  lead_type: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  vehicle_detail_link?: string | null;
  /** @nullable */
  created_at?: string | null;
}

export type GetListingsParams = {
make?: string;
model?: string;
year_min?: number;
year_max?: number;
price_min?: number;
price_max?: number;
odometer_max?: number;
limit?: number;
offset?: number;
};

export type GetListingFiltersParams = {
make?: string;
};