# lib/api-zod/src/generated/types/dealer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Used-car marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Dealer {
  dealer_id: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  postal_code?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
}