# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Used-car marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadInputLeadType } from './leadInputLeadType';

export interface LeadInput {
  /** @nullable */
  vin?: string | null;
  lead_type: LeadInputLeadType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  website?: string | null;
}