# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Used-car marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Dealer } from './dealer';

export interface Listing {
  id: string;
  vin: string;
  /** @nullable */
  dealer_id?: string | null;
  /** @nullable */
  stock_number?: string | null;
  /** @nullable */
  year?: number | null;
  /** @nullable */
  make?: string | null;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  model_number?: string | null;
  /** @nullable */
  body?: string | null;
  /** @nullable */
  transmission?: string | null;
  /** @nullable */
  series?: string | null;
  /** @nullable */
  series_detail?: string | null;
  /** @nullable */
  door_count?: number | null;
  /** @nullable */
  odometer?: number | null;
  /** @nullable */
  engine_cylinders?: number | null;
  /** @nullable */
  engine_displacement?: string | null;
  /** @nullable */
  engine?: string | null;
  /** @nullable */
  drivetrain?: string | null;
  /** @nullable */
  exterior_color?: string | null;
  /** @nullable */
  interior_color?: string | null;
  /** @nullable */
  msrp?: number | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  certified?: boolean | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  features?: string | null;
  photo_urls?: string[];
  /** @nullable */
  city_mpg?: number | null;
  /** @nullable */
  highway_mpg?: number | null;
  /** @nullable */
  vehicle_detail_link?: string | null;
  /** @nullable */
  inventory_date?: string | null;
  /** @nullable */
  photos_last_modified?: string | null;
  /** @nullable */
  age?: number | null;
  status: string;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  updated_at?: string | null;
  dealer?: Dealer;
}